refactor: rename os_call/decorator_handlers.dart → fs_decorators.dart
"decorator" is unusual in Dart; fs_decorators better describes the
VFS combinator functions (readOnly, overlay, etc.) inside the file.
