+++ lib/dart_monty_bridge.dart
@@ -27,8 +27,8 @@ export 'src/host/param.dart';
 export 'src/host/param_type.dart';
 export 'src/host/render_hint.dart';
 export 'src/host/schema.dart';
-export 'src/os_call/decorator_handlers.dart';
 export 'src/os_call/default_os_handler.dart';
+export 'src/os_call/fs_decorators.dart';
 export 'src/os_call/fs_handlers.dart';
 export 'src/os_call/os_call_exception.dart';
 export 'src/os_call/os_handlers.dart';
